expertino: move protobuf msgs to own package
This is currently a bit awkward, as the cx protobuf feature assumes
messages to be in the agent_dir.
However, for other packages it would make sense if the protobuf msgs are
separated from the agent.

Ideally the behavior of the cx protobuf feature should be changed to
accept a different relative path without too much trouble.

For now, just symlink the messages when installing to circumvent the
issue.