expertino: add agenda, redo action precondition checks
Add the concept of an agenda, a managed set of actions that are
currently available for execution to the expertino. The agenda always
looks at one equivalence class in the plan, starting at the first one
and moves ahead once all actions have been executed. A dynamic look-
ahead window is added to increase the number of potential actions to
execute. As part of this, move states to agenda items, and precondition
checks to their own facts.