expertino: add action filter for planning
The idea is to only consider a subset of all available actions for
planning. In order to determine these actions we can define a separate
planning instance that only contains actions relevant for planning.
Then the list of action names can be retrieved via a new service
request.

Note that we do not directly plan on the separate instance as this would
require us to sync all changes between two instances which is way more
offort and error-prone. Instead, we propose this more general action
filter for planning.